Added support for ASMJIT_INSTALL variable
This can be used to disable installation of asmjit files together
with the parent project. If you're building a package with CPack (make
package), you don't necessarily want asmjit headers and library to
be included in your package.

If ASMJIT_INSTALL is TRUE, everything is as usual. If it's FALSE,
nothing will be installed with make install and make package (which
is the default).