ifdef'd code to just compile on linux and some cleanup
The jitdump file is only meant for profiling with perf on linux. The code now
only compiles on linux and is disabled for other platforms.

Additionally, error handling was improved and the one use of digit separators
was removed as it requires newer C++ versions which older gcc versions do not
support.