fix gcc 8 warnings
When including the asmjit headers in a project build with
gcc 8 and -Wall -Wextra the headers trigger several compiler warnings,
which makes them hard to use (in particular in combination with -Werror).

The main problems are memcpy/memset on non-trivial objects, hidden virtual
functions, and bit arithmetic on bool (with implicit int promotion).
This commit fixes all of them by using more idiomatic C++ constructs.