Enable try mode in RA local's switchToAssignment()
This feature has been disabled for a long time so
it could be tested properly, but production didn't
reveal any issues.

When try mode is enabled the RA will try to allocate
the reassignment first to avoid possibly having to
emit code in a separate block (try mode basically
"tries" to emit code before a branch and not as a
consequence of it).