[abi] Changed signature of Xmm regs to Vec regs in x86 emitter
The purpose of this change is to minimize casting when using Vec
as a type for all vector registers in user code. In these cases
AsmJit required to cast Vec to Xmm to call emitter methods. After
this change it's no longer necessary to do that, which simplifies
user code.